+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 
 - info widget - remove margin on small table (#2727) (@dnicolson)
 - upgrade `xterm` to `v6` + style improvements
+- added option: global terminal/log window settings
 
 ### Homebridge Dependencies